eppic

+++ operations/rotate.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   rotate.c                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: dienasci <dienasci@student.42sp.org.br >   +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/01/16 19:59:17 by dienasci          #+#    #+#             */
+/*   Updated: 2022/01/16 19:59:19 by dienasci         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "../includes/push_swap.h"
 
 void	rotate_a(int **a, int silent, size_t size)
@@ -9,7 +21,7 @@ void	rotate_a(int **a, int silent, size_t size)
 		write(1, "ra\n", 4);
 	temp = (*a)[0];
 	index = 0;
-	while (index < size - 1)
+	while ((size_t)index < size - 1)
 	{
 		(*a)[index] = (*a)[index + 1];
 		index++;
